Uploads to S3

+++ packages/api/app/Controllers/Http/UploadController.js
@@ -1,11 +1,17 @@
+/* eslint-disable no-underscore-dangle */
 const Helpers = use('Helpers');
 const Upload = use('App/Models/Upload');
-const fs = require('fs').promises;
+const fs = require('fs');
+const fsp = require('fs').promises;
+
+const Logger = use('Logger');
 
 const { antl, errors, errorPayload, getTransaction } = require('../../Utils');
 
 const Config = use('Adonis/Src/Config');
 const { uploadsPath } = Config.get('upload');
+
+const Drive = use('Drive');
 
 class UploadController {
 	async index({ request }) {
@@ -26,53 +32,52 @@ class UploadController {
 		const files = request.file('files');
 		const objectInfo = meta ? JSON.parse(meta) : {};
 
-		const uploadPath = objectInfo.object
-			? `${uploadsPath}/${objectInfo.object}`
+		const folder = objectInfo.object
+			? `${uploadsPath}/${objectInfo.object}/${objectInfo.object_id}`
 			: `${uploadsPath}`;
-
-		const uploadedFiles = files.files;
 
 		let trx;
 		let uploads;
 		try {
 			const { init, commit } = getTransaction();
-			trx = await init();
-
-			await Promise.all(
-				uploadedFiles.map(async (file) => {
-					const filename = await Upload.getUniqueFileName(file, objectInfo.object);
-					await file.move(Helpers.publicPath(uploadPath), {
-						name: filename,
-					});
-				}),
-			);
-
-			if (!files.movedAll()) {
-				throw new Error('Moving files error');
-			}
-
 			const user = await auth.getUser();
 
+			trx = await init();
+
 			uploads = await Promise.all(
-				files.movedList().map((file) =>
-					user.uploads().create(
+				files.files.map(async (file) => {
+					const fileName = `${new Date().getTime()}_${file.clientName}`;
+
+					const filePath = `${Helpers.tmpPath(folder)}/${fileName}`;
+					await file.move(Helpers.tmpPath(folder), { name: fileName, overwrite: true });
+
+					const fileStream = await fs.createReadStream(filePath);
+
+					const url = await Drive.put(`${folder}/${fileName}`, fileStream, {
+						ACL: 'public-read',
+						ContentType: `${file.type}/${file.subtype}`,
+					});
+
+					fileStream._destroy();
+					Drive.disk('local').delete(filePath);
+
+					return user.uploads().create(
 						{
 							filename: file.fileName,
 							object: objectInfo.object,
 							object_id: objectInfo.object_id,
+							url,
 						},
 						trx,
-					),
-				),
+					);
+				}),
 			);
 
 			await commit();
 		} catch (error) {
 			trx.rollback();
 			await Promise.all(
-				files
-					.movedList()
-					.map((file) => fs.unlink(Helpers.publicPath(`${uploadPath}/${file.fileName}`))),
+				uploads.map((file) => Drive.delete(new URL(file.url).pathname.slice(1))),
 			);
 			return response
 				.status(400)
@@ -88,26 +93,36 @@ class UploadController {
 
 	async destroy({ params, response }) {
 		const upload = await Upload.findOrFail(params.id);
-		const uploadPath = upload.object ? `${uploadsPath}/${upload.object}` : `${uploadsPath}`;
 
-		const result = await upload.delete();
+		const file = new URL(upload.url).pathname.slice(1);
 
-		if (result) {
-			const path = Helpers.publicPath(`${uploadPath}/${upload.filename}`);
-			await fs
-				.access(path)
-				.then(() => fs.unlink(path))
-				// eslint-disable-next-line no-console
-				.catch(() => console.error('File does not exist'));
+		if (upload.url.includes('amazonaws')) {
+			if (await Drive.exists(file)) {
+				try {
+					await Drive.delete(file);
+					await upload.delete();
+				} catch (error) {
+					return response
+						.status(400)
+						.send(
+							errorPayload(
+								errors.RESOURCE_DELETED_ERROR,
+								antl('error.resource.resourceDeletedError'),
+							),
+						);
+				}
+			} else {
+				Logger.error('File does not exist.');
+			}
 		} else {
-			return response
-				.status(400)
-				.send(
-					errorPayload(
-						errors.RESOURCE_DELETED_ERROR,
-						antl('error.resource.resourceDeletedError'),
-					),
-				);
+			const localPath = upload.object ? `${uploadsPath}/${upload.object}` : `${uploadsPath}`;
+			const path = Helpers.publicPath(`${localPath}/${upload.filename}`);
+
+			await upload.delete();
+			await fsp
+				.access(path)
+				.then(() => fsp.unlink(path))
+				.catch(() => Logger.error('File does not exist'));
 		}
 
 		return response.status(200).send({ success: true });
